resolvers: move root resolvers into folders
